requirements.txt
alembic==1.15.1
aniso8601==10.0.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysqlclient==2.0.3
pycparser==2.22
PyJWT==2.10.1
python-dotenv==0.19.0
pytz==2025.2
six==1.17.0
SQLAlchemy==1.4.46
typing_extensions==4.12.2
Werkzeug==2.0.1